fix: pre-publish and post-publish hooks
