+++ packages/core/package.json
@@ -49,8 +49,8 @@
     "scripts": {
         "build": "vite build",
         "dev:iife": "concurrently \"vite build --config vite.iife.config.js --watch\" \"npx serve dist/iife -p 5174\"",
-        "prepublishOnly": "node ../../scripts/prepare-publish.js",
-        "postpublish": "node ../../scripts/restore-dev.js"
+        "prepublishOnly": "node ../../scripts/pre-publish.js",
+        "postpublish": "node ../../scripts/post-publish.js"
     },
     "dependencies": {
         "chalk": "^5.6.2",
