+++ packages/llms/package.json
@@ -43,8 +43,8 @@
     "homepage": "https://alibaba.github.io/page-agent/",
     "scripts": {
         "build": "vite build",
-        "prepublishOnly": "node ../../scripts/prepare-publish.js",
-        "postpublish": "node ../../scripts/restore-dev.js"
+        "prepublishOnly": "node ../../scripts/pre-publish.js",
+        "postpublish": "node ../../scripts/post-publish.js"
     },
     "dependencies": {
         "chalk": "^5.6.2"
